Update of ReservationController.java and Reservation.java files
Update the method for createReservation in ReservationController, both entry of reservation date and reservation time are complete.

Alternative implementation methods for reservation time have been written, an optimal one needs selection. Either using a switch or using free entry by the user and then validating it.

Collection of number of pax for reservation as well as the guest's particulars are done.

Pending: Figuring out implementation for updating a reservation and deleting reservation.